requirements/requirements.txt
# SHA1:284e5ce9a8b31c2ee7b15ed0f235259504435b85
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
asttokens==2.4.1
    # via stack-data
boto3==1.34.145
    # via -r requirements/requirements.in
botocore==1.34.145
    # via
    #   boto3
    #   s3transfer
decorator==5.1.1
    # via ipython
executing==2.0.1
    # via stack-data
ipython==8.26.0
    # via -r requirements/requirements.in
jedi==0.19.1
    # via ipython
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
matplotlib-inline==0.1.7
    # via ipython
numpy==2.0.1
    # via
    #   -r requirements/requirements.in
    #   pandas
    #   pyarrow
pandas==2.2.2
    # via -r requirements/requirements.in
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
prompt-toolkit==3.0.47
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
py4j==0.10.9.7
    # via pyspark
pyarrow==17.0.0
    # via -r requirements/requirements.in
pygments==2.18.0
    # via ipython
pyspark==3.5.1
    # via -r requirements/requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
python-dotenv==1.0.1
    # via -r requirements/requirements.in
pytz==2024.1
    # via pandas
s3transfer==0.10.2
    # via boto3
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
stack-data==0.6.3
    # via ipython
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via ipython
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via botocore
wcwidth==0.2.13
    # via prompt-toolkit
